+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './ButtonEvents';
+export * from './Button';
 export * from './CheckBox';
 export * from './DoubleSlider';
 export * from './Input';
