+++ src/stories/FancyButton/ButtonGraphics.stories.ts
@@ -1,7 +1,9 @@
 import { Graphics } from '@pixi/graphics';
-import { Button } from '../../Button';
+import { Text } from '@pixi/text';
+import { FancyButton } from '../../FancyButton';
 import { action } from '@storybook/addon-actions';
 import { argTypes, getDefaultArgs } from '../utils/argTypes';
+import { defaultTextStyle } from '../../utils/helpers/styles';
 import { centerElement } from '../../utils/helpers/resize';
 import { preloadAssets } from '../utils/loader';
 import { Container } from '@pixi/display';
@@ -9,17 +11,20 @@ import { Sprite } from '@pixi/sprite';
 import { MaskedFrame } from '../../MaskedFrame';
 
 const args = {
+    text: 'Click me!',
+    textColor: '#FFFFFF',
     color: '#A5E24D',
     hoverColor: '#FEC230',
     pressedColor: '#FE6048',
     disabledColor: '#6E6E6E',
-    width: 250,
-    height: 250,
-    padding: 10,
-    radius: 200,
-    iconSize: 250,
+    width: 350,
+    height: 350,
+    padding: 11,
+    radius: 50,
     iconOffsetX: 0,
-    iconOffsetY: 0,
+    iconOffsetY: -30,
+    textOffsetX: 0,
+    textOffsetY: 140,
     defaultOffset: 0,
     hoverOffset: -1,
     pressedOffset: 5,
@@ -28,19 +33,22 @@ const args = {
     action: action('button Event:')
 };
 
-export const UseIcon = ({
+export const UseGraphics = ({
     width,
     height,
     radius,
+    text,
     color,
     hoverColor,
     pressedColor,
     disabledColor,
     disabled,
     padding,
-    iconSize,
+    textColor,
     iconOffsetX,
     iconOffsetY,
+    textOffsetX,
+    textOffsetY,
     defaultOffset,
     hoverOffset,
     pressedOffset,
@@ -59,30 +67,38 @@ export const UseIcon = ({
 
     preloadAssets(assets).then(() =>
     {
+        const fill = Number(textColor.replace('#', '0x'));
         const target = Sprite.from(`avatar-01.png`);
 
-        target.scale.set(iconSize / target.width);
-
+        // Component usage !!!
         const icon = new MaskedFrame({
             target,
             mask: new Graphics().beginFill(0x000000).drawCircle(target.width / 2, target.height / 2, target.width / 2),
-            borderWidth: 5,
-            borderColor: 0xffffff
+            borderWidth: 10,
+            borderColor: fill
         });
 
         // Component usage !!!
-        const button = new Button({
+        const button = new FancyButton({
             defaultView: new Graphics().beginFill(color).drawRoundedRect(0, 0, width, height, radius),
             hoverView: new Graphics().beginFill(hoverColor).drawRoundedRect(0, 0, width, height, radius),
             pressedView: new Graphics().beginFill(pressedColor).drawRoundedRect(0, 0, width, height, radius),
             disabledView: new Graphics().beginFill(disabledColor).drawRoundedRect(0, 0, width, height, radius),
             icon,
+            text: new Text(text, {
+                ...defaultTextStyle,
+                fill
+            }),
             padding,
             offset: {
                 default: { y: defaultOffset },
                 hover: { y: hoverOffset },
                 pressed: { y: pressedOffset },
                 disabled: { y: disabledOffset }
+            },
+            textOffset: {
+                x: textOffsetX,
+                y: textOffsetY
             },
             iconOffset: {
                 x: iconOffsetX,
@@ -113,7 +129,7 @@ export const UseIcon = ({
 };
 
 export default {
-    title: 'Components/Button/Use Icon',
+    title: 'Components/FancyButton/Use Graphics',
     argTypes: argTypes(args),
     args: getDefaultArgs(args)
 };
