+++ src/stories/FancyButton/ButtonIcon.stories.ts
@@ -1,9 +1,7 @@
 import { Graphics } from '@pixi/graphics';
-import { Text } from '@pixi/text';
-import { Button } from '../../Button';
+import { FancyButton } from '../../FancyButton';
 import { action } from '@storybook/addon-actions';
 import { argTypes, getDefaultArgs } from '../utils/argTypes';
-import { defaultTextStyle } from '../../utils/helpers/styles';
 import { centerElement } from '../../utils/helpers/resize';
 import { preloadAssets } from '../utils/loader';
 import { Container } from '@pixi/display';
@@ -11,21 +9,17 @@ import { Sprite } from '@pixi/sprite';
 import { MaskedFrame } from '../../MaskedFrame';
 
 const args = {
-    text: 'Click me!',
-    textColor: '#FFFFFF',
     color: '#A5E24D',
     hoverColor: '#FEC230',
     pressedColor: '#FE6048',
     disabledColor: '#6E6E6E',
-    width: 350,
-    height: 200,
-    padding: 11,
-    radius: 50,
-    iconSize: 100,
+    width: 250,
+    height: 250,
+    padding: 10,
+    radius: 200,
+    iconSize: 250,
     iconOffsetX: 0,
-    iconOffsetY: -30,
-    textOffsetX: 0,
-    textOffsetY: 50,
+    iconOffsetY: 0,
     defaultOffset: 0,
     hoverOffset: -1,
     pressedOffset: 5,
@@ -34,23 +28,19 @@ const args = {
     action: action('button Event:')
 };
 
-export const UseGraphics = ({
+export const UseIcon = ({
     width,
     height,
     radius,
-    text,
     color,
     hoverColor,
     pressedColor,
     disabledColor,
     disabled,
     padding,
-    textColor,
     iconSize,
     iconOffsetX,
     iconOffsetY,
-    textOffsetX,
-    textOffsetY,
     defaultOffset,
     hoverOffset,
     pressedOffset,
@@ -69,40 +59,28 @@ export const UseGraphics = ({
 
     preloadAssets(assets).then(() =>
     {
-        const fill = Number(textColor.replace('#', '0x'));
         const target = Sprite.from(`avatar-01.png`);
 
-        // Component usage !!!
         const icon = new MaskedFrame({
             target,
             mask: new Graphics().beginFill(0x000000).drawCircle(target.width / 2, target.height / 2, target.width / 2),
-            borderWidth: 10,
-            borderColor: fill
+            borderWidth: 5,
+            borderColor: 0xffffff
         });
 
-        icon.scale.set(iconSize / target.width);
-
         // Component usage !!!
-        const button = new Button({
+        const button = new FancyButton({
             defaultView: new Graphics().beginFill(color).drawRoundedRect(0, 0, width, height, radius),
             hoverView: new Graphics().beginFill(hoverColor).drawRoundedRect(0, 0, width, height, radius),
             pressedView: new Graphics().beginFill(pressedColor).drawRoundedRect(0, 0, width, height, radius),
             disabledView: new Graphics().beginFill(disabledColor).drawRoundedRect(0, 0, width, height, radius),
             icon,
-            text: new Text(text, {
-                ...defaultTextStyle,
-                fill
-            }),
             padding,
             offset: {
                 default: { y: defaultOffset },
                 hover: { y: hoverOffset },
                 pressed: { y: pressedOffset },
                 disabled: { y: disabledOffset }
-            },
-            textOffset: {
-                x: textOffsetX,
-                y: textOffsetY
             },
             iconOffset: {
                 x: iconOffsetX,
@@ -133,7 +111,7 @@ export const UseGraphics = ({
 };
 
 export default {
-    title: 'Components/Button/Use Graphics',
+    title: 'Components/FancyButton/Use Icon',
     argTypes: argTypes(args),
     args: getDefaultArgs(args)
 };
